Module hydra_dx_math::ema

Functions

balance_weighted_average

Calculate a weighted average for the given balances. prev is the previous oracle value, incoming is the new value to integrate. weight is how much weight to give the new value.

calculate_new_by_integrating_incoming

Calculate the new oracle values by integrating incoming values with the previous oracle. Uses a weighted average based on the smoothing factor.

exp_smoothing

Calculate the smoothing factor for a period from a given combination of original smoothing factor and iterations by exponentiating the complement by the iterations.

iterated_balance_ema

Calculate the iterated exponential moving average for the given balances. iterations is the number of iterations of the EMA to calculate. prev is the previous oracle value, incoming is the new value to integrate. smoothing is the smoothing factor of the EMA.

iterated_liquidity_ema

Calculate the iterated exponential moving average for the given balances. iterations is the number of iterations of the EMA to calculate. prev is the previous oracle value, incoming is the new value to integrate. smoothing is the smoothing factor of the EMA.

iterated_price_ema

Calculate the iterated exponential moving average for the given prices. iterations is the number of iterations of the EMA to calculate. prev is the previous oracle value, incoming is the new value to integrate. smoothing is the smoothing factor of the EMA.

iterated_volume_ema

Calculate the iterated exponential moving average for the givenEmaVolumes. iterations is the number of iterations of the EMA to calculate. prev is the previous oracle value; the incoming value is always zero. smoothing is the smoothing factor of the EMA.

liquidity_weighted_average

Calculate a weighted average for the givenEmaLiquidity values. prev is the previous oracle value, incoming is the new value to integrate. weight is how much weight to give the new value.

price_weighted_average

Calculate a weighted average for the given prices. prev is the previous oracle value, incoming is the new value to integrate. weight is how much weight to give the new value.

smoothing_from_period

Calculates smoothing factor alpha for an exponential moving average based on period: alpha = 2 / (period + 1). It leads to the “center of mass” of the EMA corresponding to be the “center of mass” of a period-length SMA.

update_outdated_to_current

Calculate the current oracle values from the outdated and update_with values using the smoothing factor with the old values being iterations out of date.

volume_weighted_average

Calculate a weighted average for the givenEmaVolumes. prev is the previous oracle value, incoming is the new value to integrate. weight is how much weight to give the new value.

Type Definitions

EmaLiquidity

EmaPrice

EmaPrice is a rational number represented by a u128 for both numerator and denominator.

EmaVolume