Module hydra_dx_math::lbp

Macros

to_balance_from_fixed

to_fixed_balance

Functions

calculate_in_given_out

Calculating buying price given reserve of selling asset and reserve of buying asset. Formula :

calculate_linear_weights

Calculating weight at any given block in an interval using linear interpolation.

calculate_out_given_in

Calculating selling price given reserve of selling asset and reserve of buying asset.

calculate_spot_price

Calculating spot price given reserve of selling asset and reserve of buying asset. Formula : BUY_RESERVE * AMOUNT / SELL_RESERVE