Module hydra_dx_math::liquidity_mining

Functions

calculate_accumulated_rps

This function calculate and return reward per share or error.

calculate_global_pool_reward_per_period

This function return GlobalPool’s reward per one period or error. Reward per period is capped by max_reward_per_period.

calculate_global_pool_shares

This function calculate liq. pool’s shares amount in GlobalPool or error.

calculate_loyalty_multiplier

This function calculate loyalty multiplier or error.

calculate_reward

General formula to calculate reward. Usage depends on type of rps and shares used for calculations

calculate_user_reward

This function calculate and return (user_rewards, unclaimable_rewards).

calculate_valued_shares

This function calculate account’s valued shares[Balance] or error.