Module hydra_dx_math::stableswap::math

Modules

two_asset_pool_math

Stableswap/curve math reduced to two assets.

Functions

calculate_add_liquidity_shares

Calculate shares amount after liquidity is added to the pool.

calculate_in_given_out

Calculating amount to be sent to the pool given the amount to be received from the pool and both reserves. N - number of iterations to use for Newton’s formula ( it should be >=1 otherwise it wont converge at all and will always fail N_Y - number of iterations to use for Newton’s formula to calculate reserve Y ( it should be >=1 otherwise it wont converge at all and will always fail

calculate_out_given_in

Calculating amount to be received from the pool given the amount to be sent to the pool and both reserves. N - number of iterations to use for Newton’s formula to calculate parameter D ( it should be >=1 otherwise it wont converge at all and will always fail N_Y - number of iterations to use for Newton’s formula to calculate reserve Y ( it should be >=1 otherwise it wont converge at all and will always fail

calculate_remove_liquidity_amounts

Given amount of shares and asset reserves, calculate corresponding amounts of each asset to be withdrawn.