Function hydra_dx_math::stableswap::math::calculate_out_given_in

pub fn calculate_out_given_in<const N: u8, const N_Y: u8>(
    reserve_in: Balance, 
    reserve_out: Balance, 
    amount_in: Balance, 
    amplification: Balance, 
    precision: Balance
) -> Option<Balance>

Calculating amount to be received from the pool given the amount to be sent to the pool and both reserves. N - number of iterations to use for Newton’s formula to calculate parameter D ( it should be >=1 otherwise it wont converge at all and will always fail N_Y - number of iterations to use for Newton’s formula to calculate reserve Y ( it should be >=1 otherwise it wont converge at all and will always fail